Moved derived models into model_list
